+++ fs/fjs/module.f.ts
@@ -3,16 +3,15 @@
  *
  * @module
  */
-import { fromIo, type Io } from '../io/module.f.ts'
+import { fromIo, runProgram, type Io } from '../io/module.f.ts'
 import { compile } from '../djs/module.f.ts'
 import { main as testMain } from '../dev/tf/module.f.ts'
 import { main as casMain } from '../cas/module.f.ts'
-import { import_, type NodeProgram, type NodeProgramOptions } from '../types/effects/node/module.f.ts'
+import { import_, type NodeProgram } from '../types/effects/node/module.f.ts'
 
 export const main = async(io: Io): Promise<number> => {
     const { error } = io.console
     const { process } = io
-    const { env } = process
     const [command, ...rest] = process.argv.slice(2)
     const eRun = fromIo(io)
     switch (command) {
@@ -28,10 +27,9 @@ export const main = async(io: Io): Promise<number> => {
         case 'run':
         case 'r':
             const [file, ...args] = rest
-            return eRun(import_(file).step(result => {
-                if (result[0] === 'error') { throw result[1] }
-                const options: NodeProgramOptions = { args, env }
-                return (result[1].default as NodeProgram)(options)
+            return runProgram(io)(args)(o => import_(file).step(([s, v]) => {
+                if (s === 'error') { throw v }
+                return (v.default as NodeProgram)(o)
             }))
         case undefined:
             error('Error: command is required')