+++ fs/types/effects/node/virtual/module.f.ts
@@ -7,6 +7,7 @@ import { todo } from '../../../../dev/module.f.ts'
 import { parse } from '../../../../path/module.f.ts'
 import { isVec, type Vec } from '../../../bit_vec/module.f.ts'
 import { error, ok } from '../../../result/module.f.ts'
+import { decodeUtf8, fromVec } from '../../../uint8array/module.f.ts'
 import { run, type MemOperationMap, type RunInstance } from '../../mock/module.f.ts'
 import type { Dirent, IoResult, NodeOp } from '../module.f.ts'
 
@@ -167,6 +168,10 @@ const map: MemOperationMap<NodeOp, State> = {
         let result
         try { result = ok(f()) } catch (e) { result = error(e) }
         return [state, { result, duration: 0 }]
+    },
+    write: (state, stream, data) => {
+        const s = decodeUtf8(fromVec(data))
+        return [{ ...state, [stream]: `${state[stream]}${s}` }, undefined] as const
     },
 }